requirements.txt
Flask==2.2.2
grpc==1.0.0
grpcio==1.50.0
matplotlib==3.6.3
mysql_connector_repackaged==0.3.1
numpy==1.23.4
protobuf==4.21.12
requests==2.28.1
spyne==2.14.0
